requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
scipy>=1.10
statsmodels>=0.14
matplotlib>=3.7
seaborn>=0.13
joblib>=1.3
fastapi>=0.110
uvicorn[standard]>=0.29
pydantic>=2.0
pika>=1.3
psycopg2-binary>=2.9
